ft_transcendance/django/requirements.txt
channels==3.0.5
channels-redis==4.1.0
psycopg2-binary==2.9.9
Django==4.2.4
djangorestframework==3.15.1
pillow==10.3.0
pyOpenSSL==24.1.0
redis==5.0.4
djangorestframework-simplejwt
bcrypt
django-cors-headers
Pillow
uvicorn
requests
websockets
pytz